fix(uptime): Centralize assertion normalization in UptimeAssertionsField
Move normalizeAssertion from uptimeAlertForm.tsx to field.tsx and apply
it via the getData prop on FormField. This ensures assertion values are
normalized before submission for both:
- Alerts form (/issues/alerts/new/uptime/)
- Detectors form (/monitors/new/)

The normalization handles NaN values (from cleared inputs) by defaulting
to 200, and clamps values to the valid HTTP status code range (100-599).
