+++ static/app/views/alerts/rules/uptime/assertions/field.tsx
@@ -1,8 +1,37 @@
 import type {FormFieldProps} from 'sentry/components/forms/formField';
 import FormField from 'sentry/components/forms/formField';
 import {uniqueId} from 'sentry/utils/guid';
+import type {Assertion, Op} from 'sentry/views/alerts/rules/uptime/types';
 
 import {AssertionOpGroup} from './opGroup';
+
+/**
+ * Recursively normalizes assertion values to ensure they are valid before submission.
+ * Handles NaN values (from cleared inputs) and clamps to valid HTTP status code range.
+ */
+export function normalizeAssertion(op: Op): Op {
+  switch (op.op) {
+    case 'status_code_check':
+      return {
+        ...op,
+        // Default to 200 if NaN (e.g., user cleared input and submitted without blur)
+        value: isNaN(op.value) ? 200 : Math.max(100, Math.min(599, op.value)),
+      };
+    case 'and':
+    case 'or':
+      return {
+        ...op,
+        children: op.children.map(normalizeAssertion),
+      };
+    case 'not':
+      return {
+        ...op,
+        operand: normalizeAssertion(op.operand),
+      };
+    default:
+      return op;
+  }
+}
 
 // XXX(epurkhiser): The types of the FormField render props are absolutely
 // abysmal, so we're leaving this untyped for now.
@@ -29,6 +58,9 @@ export function UptimeAssertionsField(props: Omit<FormFieldProps, 'children'>) {
       defaultValue={{root: {op: 'and', children: [], id: uniqueId()}}}
       {...props}
       flexibleControlStateSize
+      getData={(data: Assertion | undefined) =>
+        data?.root ? {root: normalizeAssertion(data.root)} : data
+      }
     >
       {({ref: _ref, ...fieldProps}) => <UptimeAssertionsControl {...fieldProps} />}
     </FormField>
